B4
v-3.0-p1-t1
---
1. Data type => "numeric"

+++ db/migrate/20130905050910_add_column_optimal_percentage_to_admin.rb
@@ -2,7 +2,8 @@ class AddColumnOptimalPercentageToAdmin < ActiveRecord::Migration
   def up
     
     # REF http://stackoverflow.com/questions/4384284/rails-generate-model-fieldtype-what-are-the-options-for-fieldtype answered Dec 8 '10 at 5:02
-    add_column :admins, :optimal_percentage, :float
+    #add_column :admins, :optimal_percentage, :float
+    add_column :admins, :optimal_percentage, :numeric
     
   end
